LUN-130: Move VocabularyList Modal to separate file
