+++ src/components/ImageCarousel.tsx
@@ -1,20 +1,24 @@
 import React from 'react'
-import { useWindowDimensions, View } from 'react-native'
+import { useWindowDimensions } from 'react-native'
 import { ImagesType } from '../constants/endpoints'
 import styled from 'styled-components/native'
 import ImageViewer from 'react-native-image-zoom-viewer'
 import { COLORS } from '../constants/colors'
 import { Pagination } from 'react-native-snap-carousel'
 
+const ImageView = styled.View<{ viewportHeight: number }>`
+  height: ${props => 0.4 * props.viewportHeight}px;
+`
+
 const StyledImage = styled.Image`
   height: 100%;
 `
 
-const PaginationView = styled.View`
+const PaginationView = styled.View<{ viewportHeight: number }>`
   position: absolute;
   left: 0;
   right: 0;
-  top: -10;
+  top: ${props => props.viewportHeight * -0.01}px;
   background-color: transparent;
 `
 
@@ -29,14 +33,13 @@ interface ImageUrlType {
 const ImageCarousel = ({ images }: ImageCarouselPropsType) => {
   const imagesUrls: ImageUrlType[] = []
   images.forEach(it => imagesUrls.push({ url: it.image }))
-  const { height: viewportHeight, width: viewportWidth } = useWindowDimensions()
-  console.log(viewportWidth)
+  const { height: viewportHeight } = useWindowDimensions()
 
   const renderIndicator = (currentIndex?: number, allSize?: number) => {
     return !currentIndex || !allSize ? (
       <></>
     ) : (
-      <PaginationView>
+      <PaginationView viewportHeight={viewportHeight}>
         <Pagination
           activeDotIndex={currentIndex - 1}
           dotsLength={allSize}
@@ -52,14 +55,14 @@ const ImageCarousel = ({ images }: ImageCarouselPropsType) => {
   }
 
   return (
-    <View style={{ height: 0.4 * viewportHeight, width: 1.5 * viewportWidth, marginLeft: -105 }}>
+    <ImageView viewportHeight={viewportHeight}>
       <ImageViewer
         imageUrls={imagesUrls}
         renderImage={renderItem}
         renderIndicator={renderIndicator}
         backgroundColor={COLORS.lunesWhite}
       />
-    </View>
+    </ImageView>
   )
 }
